notebook-app/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
google-generativeai
jinja2
